backend/functions/nlp_engine/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
google-cloud-pubsub==2.*
openai==1.*
